feat: introduce prefix-based layout
This is an important change for the entire playbook that is intended
to solve various code organization problems: ordering, navigation, and
extensibility.

Proposed solution defines two-level level hierarchy for each role in
this project.

First level directory must have [a-z]a- prefix, where each character
represents provisioning flow. A prefix aa- is reserved for global
namespace that are shared across different roles.

Second level directory must be a role. A prefix [a-z]a- is reserved
for shared namespace that is used in any role defined later.