feat(foundation): implement user layout management
This commit introduces a successor to common_* layout action plugins.

New plugins do not share any code with previous approaches. Now it is
possible to have a full user layout resolution flow thanks to the
dedicated operation analyzer. The verdict can be used then to wipe and
create links or simply filter and return result in case of lookup
plugin.