feat(foundation): add foundation.user.env action plugin
This action plugin introduces a new declarative way to manage user
environmental variables. It parses pam_env.conf and applies new changes.

The parser supports commentaries (top/inline) and performs a path
substitution. This way, every path is converted into the user-agnostic
expression that gets evaluated by the pam_env itself.

The plugin requires that the variables to be part of a section. This
visually separates them and makes further manual editing easier.

It is a very powerful tool that can be used together which complements
user layout action plugins.