feat(foundation): add foundation.persist.* action plugins
It is mostly an adaptation of previous common_persist to a new
foundation architecture. It also allows to retrieve user key metadata
directly from the provisioned machine.

The plugins now require variable-based usage and use shared modules.
Persist unpacking introduces a guard mechanism that would skip the
process entirely if some of the keys are not available. It is
particularly useful for simple backup-not-exists flow without nested
task blocks.