feat(desktop)!: add intel graphics stack support
This changes the hard nvidia dependency, and allows to specify
alternative drivers and Xorg configuration. At this point of time,
it includes nvidia and intel GPUs, but this can be changed in the
future.

BREAKING CHANGE: `ga_stack` variable needs to be defined in the
inventory. See the example inventory for more information.