feat(prelude)!: use foundation.storage.partition plugin
This change replaces apart command-line utility with native
foundation.storage.partition plugin.

While the apart was a solid choice served, it is still a command-line
utility that poorly integrates with the Ansible ecosystem. Printing
the raw stdout back to the user may not always be the best option.

The action plugin makes use of new declarative syntax optimized for
ansible inventory instead of the command-line arguments. It also
allows to configure mount and filesystem options in one place, as well
as generate the fstab entries without patching genfstab results.

Swap partitions can now be properly mounted and reflected in the fstab.

BREAKING CHANGE:
1) `bb_schema` syntax has changed. Follow the
`foundation.storage.partition` documentation in the README for more
details.
2) `bc_mountopts_root` and `bc_mountopts_efi` are no longer used. The
same functionality is now integrated to the `bb_schema`.
3) fstab generation is now a part of the `bb-partition` role instead of
the `bc-bootstrap`.