feat(bootloader): inherit kernel params for root partition from fstab
The disk layout variable `bb_schema` has always been misaligned with
the kernel parameters read from `dc_cmdline`. This change allows to
omit `rootfstype` and `rootflags` from `dc_cmdline`. If this is the
case, then these parameters will be constructed directly from the
existing `/etc/fstab`.

No breaking changes are introduced. If `rootfstype` and `rootflags`
are present in the `dc_cmdline`, they will not be overwritten.