feat(apps): add miniserve configuration
This commit adds basic miniserve configuration that might be useful
to omit on each run (because it is meant to "just" work).
This options still can be overwritten via the cli if needed.

It also fixes a problem with default interfaces list that may cause
miniserve to refuse start when the IPv6 stack is disabled.